urlopy i timer

+++ app/src/main/java/com/example/worktracker/Holiday.java
@@ -21,7 +21,7 @@ public class Holiday {
         return id;
     }
 
-    public float getLiczbaDni() {
+    public int getLiczbaDni() {
         return liczbaDni;
     }
